requirements.txt
backcall==0.1.0
decorator==4.4.0
Django==2.1.7
django-braces==1.13.0
django-crispy-forms==1.7.2
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.9
psycopg2-binary==2.7.7
ptyprocess==0.6.0
Pygments==2.3.1
pytz==2018.9
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
